Woot! Meditated every day in January 2019 and hit my first goal.
Overall I think 45 minutes is a great duration. Some things I've learned...

I started with a traditional zafu and zabuton, but after a few days of having some leg issues, switched to a seiza bench. That has greatly improved the quality of my meditation.

Focus is fleeting, but conditioning the mental response to reassert focus when it is recognized as lost is a very powerful technique. It has improved my ability to retain focus in all contexts of life outside of meditation time.

Dedicating the merit of the meditation is important! Setting the intention and generating motivation for the why of meditation is also important! Just because meditation can be a secular activity doesn't mean we should divorce it of a greater purpose than only benefitting ourselves. Meditation fundamentally is a practice that helps me be of better service to others. Any benefit to myself is a second-order effect. I think that is the most important point to focus on for anyone interested in meditation.